Added index files and restructured imports
