+++ src/api/index.ts
@@ -1,0 +1,7 @@
+import fetchRecipe from "./fetchRecipe";
+import fetchRecipes from "./fetchRecipes";
+import createRecipe from "./createRecipe";
+import modifyRecipe from "./modifyRecipe";
+import deleteRecipe from "./deleteRecipe";
+
+export { fetchRecipe, fetchRecipes, createRecipe, modifyRecipe, deleteRecipe }
