+++ src/context/index.ts
@@ -1,0 +1,3 @@
+import RecipeContext from "./recipeContext";
+
+export { RecipeContext }
